Split Mapper into smaller component files
